Refactoring, move ImageUtils
