+++ android/capacitor/src/main/java/com/getcapacitor/plugin/camera/ImageUtils.java
